backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
google-genai==1.0.0
aiofiles==24.1.0
Pillow==10.4.0
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.3
